Move the menu to separate component
